Add more unit test coverage for CarbonUtils class
